add capture port 80 packet

+++ app/src/main/java/eu/faircode/netguard/ActivityMain.java
@@ -784,7 +784,7 @@ public class ActivityMain extends AppCompatActivity implements SharedPreferences
                             if (pcap_file_tcp.exists() && !pcap_file_tcp.delete())
                                 Log.w(TAG, "Delete PCAP failed");
                         }
-                        /*if (prefs.getBoolean("Other", false)) {
+                        if (prefs.getBoolean("Other", false)) {
                             SinkholeService.setPcapother(null);
                             if (pcap_file_other.exists() && !pcap_file_other.delete())
                                 Log.w(TAG, "Delete PCAP failed");
@@ -793,7 +793,7 @@ public class ActivityMain extends AppCompatActivity implements SharedPreferences
                         else {
                             if (pcap_file_other.exists() && !pcap_file_other.delete())
                                 Log.w(TAG, "Delete PCAP failed");
-                        }*/
+                        }
                         return null;
                     }
